Rename JSONAppReducers
Rename JSONAppReducers => JSONSingleContainerReducers

close DCOS-14526
