+++ plugins/services/src/js/reducers/JSONSingleContainerReducers.js
